pyproject.toml
[tool.poetry]
name = "pythonproject62"
version = "0.1.0"
description = ""
authors = ["Borodin Dmitriy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
